Connected flask w react frontend
