--- event-backend/__init__.py
@@ -1,2 +1,0 @@
-import flask
-app = flask.Flask(__name__)
