+++ event-backend/key.py
@@ -1,0 +1,5 @@
+api_key = '8leDUNiLGu4Ks1xsl6TNcOBaA'
+api_key_secret = 'uGgkJBmwN0aFfNbkgcqRFmbnSdkG7wuSGLGu6FdZ5fd4bRN2qw'
+bearer_token = 'AAAAAAAAAAAAAAAAAAAAAGSjJwEAAAAAZKYgUhop%2FJMRhXyYx%2Fs83E8z%2FF0%3DC6Nj3ofKGmLsqC1pndc9LC829htwc93zg7BTdg7PVAknOUMuU9'
+access_token = '1067244479294377985-IDTBYA4wyHXN4ci45rM2TneHwxgmcA'
+access_token_secret = 'Wa00KzBrd5ZG2XnfqopqITKdlXOI7l9iMB2ptXleNnYhn'
